requirements.txt
requests 
python-dotenv
firebase-admin    
aiohttp 